Hjemmeside/requirements.txt
adal==1.2.7
antlr4-python3-runtime==4.13.1
applicationinsights==0.11.10
argcomplete==3.1.6
azure-core==1.30.2
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-MySQLdb==2.0.0
fonttools==4.51.0
humanfriendly==10.0
idna==3.7
invoke==2.2.0
isodate==0.6.1
itsdangerous==2.2.0
javaproperties==0.5.2
Jinja2==3.1.3
jmespath==1.0.1
jsondiff==2.0.0
kiwisolver==1.4.5
knack==0.11.0
MarkupSafe==2.1.5
matplotlib==3.8.4
msal==1.28.0
msal-extensions==1.2.0b1
msrest==0.7.1
msrestazure==0.6.4.post1
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
paramiko==3.4.0
pillow==10.3.0
pkginfo==1.10.0
portalocker==2.8.2
psutil==5.9.8
pycomposefile==0.0.31
pycparser==2.22
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
scp==0.13.6
semver==2.13.0
six==1.16.0
sshtunnel==0.1.5
tabulate==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
websocket-client==1.3.3
Werkzeug==3.0.2
wrapt==1.16.0
xmltodict==0.13.0
